refactor: move reportable-artifact-evidence checks to execution_contract
Continues eliminating core->runner behavioral edges. Like
is_remote_runner_artifact_path, `is_reportable_artifact_evidence_path` and
`reportable_artifact_evidence_path` were runner-module functions but their
logic is entirely contract-driven: a runner-artifact-scheme check, a
metadata-only-ref check, a relative-path check, and a local fs stat — all
core concerns (EXECUTION_CONTRACT lives in core).

Reimplemented both in core's execution_contract module (the retrievable-token
check becomes a direct `strip_runner_artifact_scheme(path).is_some()` contract
call, dropping the RemoteArtifactToken dependency), and repointed the 3
external-core callers (observation/bundle, extension/trace/report,
extension/bench/report). Re-exported from runner::evidence::tokens so
runner-internal call sites resolve unchanged.

Reverse-edge file count: 36 -> 33.

Verified: cargo check clean for -p homeboy-core --lib and -p homeboy-cli --lib.
