+++ crates/homeboy-core/src/execution_contract.rs
@@ -148,6 +148,28 @@ pub fn is_remote_runner_artifact_path(path: &str) -> bool {
     EXECUTION_CONTRACT.artifacts.is_runner_artifact_ref(path)
 }
 
+/// Whether an artifact path is worth reporting as retrievable evidence: a
+/// retrievable runner-artifact token, a metadata-only ref, a relative path, or
+/// a real local file/dir. Contract-driven classification (not runner behavior),
+/// so it lives in core.
+pub fn is_reportable_artifact_evidence_path(path: &str) -> bool {
+    EXECUTION_CONTRACT
+        .artifacts
+        .strip_runner_artifact_scheme(path)
+        .is_some()
+        || EXECUTION_CONTRACT.artifacts.is_metadata_only_ref(path)
+        || !std::path::Path::new(path).is_absolute()
+        || std::fs::metadata(path)
+            .map(|metadata| metadata.is_file() || metadata.is_dir())
+            .unwrap_or(false)
+}
+
+/// The path itself when it is reportable artifact evidence, else `None`.
+pub fn reportable_artifact_evidence_path(path: Option<&String>) -> Option<String> {
+    path.filter(|path| is_reportable_artifact_evidence_path(path))
+        .cloned()
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
