+++ crates/homeboy-core/src/extension/trace/report.rs
@@ -16,8 +16,8 @@ use super::span_summary::{
     TraceSpanSummaryOutput,
 };
 use crate::engine::detail_output::{bounded_items, DEFAULT_DETAIL_ITEM_LIMIT};
+use crate::execution_contract::is_reportable_artifact_evidence_path;
 use crate::rig::RigStateSnapshot;
-use crate::runner::is_reportable_artifact_evidence_path;
 
 pub use aggregate_types::*;
 pub use builders::*;
